Adding use-case to build filters to get issues
