+++ src/core/application/use-cases/issues/get-issues-by-filters.use-case.ts
@@ -7,17 +7,18 @@ import { GetIssuesByFiltersDto } from '@/core/infrastructure/http/dtos/get-issue
 import { PaginationDto } from '@/core/infrastructure/http/dtos/pagination.dto';
 import { IUseCase } from '@/shared/domain/interfaces/use-case.interface';
 import { Inject, Injectable } from '@nestjs/common';
+import { BuildFilterUseCase } from './build-filter.use-case';
 
 @Injectable()
 export class GetIssuesByFiltersUseCase implements IUseCase {
     constructor(
         @Inject(ISSUES_SERVICE_TOKEN)
         private readonly issuesService: IIssuesService,
+
+        private readonly buildFilterUseCase: BuildFilterUseCase,
     ) {}
 
-    async execute(
-        filters: GetIssuesByFiltersDto,
-    ): Promise<{
+    async execute(filters: GetIssuesByFiltersDto): Promise<{
         data: IssuesEntity[];
         pagination: {
             page: number;
@@ -26,7 +27,7 @@ export class GetIssuesByFiltersUseCase implements IUseCase {
             totalPages: number;
         };
     }> {
-        const filter = this.buildFilter(filters);
+        const filter = await this.buildFilterUseCase.execute(filters);
         const paginationOptions = this.buildPaginationOptions(filters);
 
         const [data, total] = await Promise.all([
@@ -43,48 +44,6 @@ export class GetIssuesByFiltersUseCase implements IUseCase {
                 totalPages: Math.ceil(total / (filters.limit || 100)),
             },
         };
-    }
-
-    private buildFilter(filters: GetIssuesByFiltersDto): any {
-        const filter: any = {};
-
-        if (filters.title) {
-            filter['title'] = { $regex: filters.title, $options: 'i' };
-        }
-
-        const exactMatchFields = [
-            'severity',
-            'category',
-            'status',
-            'organizationId',
-            'filePath',
-        ];
-        exactMatchFields.forEach((field) => {
-            if (filters[field]) {
-                filter[field] = filters[field];
-            }
-        });
-
-        if (filters.repositoryName) {
-            filter['repository.name'] = {
-                $regex: filters.repositoryName,
-                $options: 'i',
-            };
-        }
-
-        if (filters.beforeAt || filters.afterAt) {
-            filter['createdAt'] = {};
-
-            if (filters.beforeAt) {
-                filter['createdAt'].$lt = new Date(filters.beforeAt);
-            }
-
-            if (filters.afterAt) {
-                filter['createdAt'].$gt = new Date(filters.afterAt);
-            }
-        }
-
-        return filter;
     }
 
     private buildPaginationOptions(options: PaginationDto) {
